Retro std::size_t getUnitsSize() to int

+++ src/player/brains/missions/cPlayerBrainMissionKindDeathHand.cpp
@@ -54,14 +54,14 @@ bool cPlayerBrainMissionKindDeathHand::think_SelectTarget()
 
     if (target < 0) {
         // find any unit to attack instead
-        for (size_t i = 0; i < game.m_gameObjectsContext->getUnitsSize(); i++) {
+        for (int i = 0; i < game.m_gameObjectsContext->getUnitsSize(); i++) {
             cUnit *pUnit = game.m_gameObjectsContext->getUnit(i);
             if (!pUnit->isValid()) continue;
             if (pUnit->getPlayer() == player) continue; // skip self
             if (pUnit->getPlayer()->isSameTeamAs(player)) continue; // skip allies and self
             if (!game.m_gameObjectsContext->getMap().isVisible(pUnit->getCell(), player)) continue; // skip non visible targets
             // enemy unit
-            target = static_cast<int>(i);
+            target = i;
             if (RNG::rnd(100) < 5) {
                 break; // this way we kind of have randomly another target...
             }
